Move problem report metadata to module
